login, logout, signup, sessionid, etc

+++ djangoProject/base/views.py
@@ -1,14 +1,86 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from .models import Room, Topic
+from django.contrib.auth.models import User
 from .forms import RoomForm
+from django.db.models import Q
+from django.contrib import messages     #flash messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 # Create your views here.
 
 
-def home(request):
-    rooms = Room.objects.all()
-    topics = Topic.objects.all()
+def loginPage(request):
 
-    context = {'rooms': rooms, 'topics': topics}
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+          user = User.objects.get(username=username)
+
+        except:
+            messages.error(request, "user does not exist.")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+
+        else:
+            messages.error(request, "username or password does not exist.")
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+          user = form.save(commit=False)    # creates user object but doesn't save to DB yet
+          user.username = user.username.lower()   #converts username to lowercase
+          user.save()
+          login(request, user)
+          return redirect('home')
+        
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+
+    return render(request, 'base/login_register.html', {'form':form})
+
+
+
+def home(request):
+
+    #search functionality 
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q) |
+        Q(host__username__icontains=q)
+        )
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -16,6 +88,8 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')   #redirect back to the login page
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -28,9 +102,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login') 
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)   #this is how we get our form prefilled for update
+    
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here!")
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -41,8 +120,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login') 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here!")
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
